+++ sandbox.cjs
@@ -1,7 +1,13 @@
 const { execSync } = require('child_process');
 const { URL } = require('url');
 
+const CADENZA_URL_PARAM = '--cadenza-url';
 const DEFAULT_CADENZA_URL = 'http://localhost:8080/cadenza';
-const cadenzaUrl = new URL((process.argv[2] === '--cadenza-url' && process.argv[3]) || process.env.CADENZA_URL || DEFAULT_CADENZA_URL);
+
+const index = process.argv.indexOf(CADENZA_URL_PARAM);
+const cadenzaUrlFromArgs = index !== -1
+    ? process.argv[index + 1]
+    : process.argv.find(arg => arg.startsWith(CADENZA_URL_PARAM + '='))?.slice(CADENZA_URL_PARAM.length + 1);
+const cadenzaUrl = new URL(cadenzaUrlFromArgs || process.env.CADENZA_URL || DEFAULT_CADENZA_URL);
 
 execSync(`npx --yes http-server -a localhost -c-1 -d false --proxy ${cadenzaUrl.origin} --proxy-options.headers.Origin ${cadenzaUrl.origin} -o /sandbox.html${cadenzaUrl.pathname !== '/' ? `?contextPath=${cadenzaUrl.pathname}` : ''}`, { stdio: 'inherit' });